Fin ajout livre, clique State et Indicator Btn, sauvegarde fichier

+++ app/src/main/java/com/example/mystilib/events/AddBook.java
@@ -5,5 +5,5 @@ public class AddBook {
     public String groupname;
     public int state;
     public int indicator;
-    public int stars;
+    public float stars;
 }
